Move the text and colour sections into their own file
ControlPanelSections had fourteen composables; the six about text and colour go
to ControlPanelTextSections.kt.
